fix(frontend): renames eth token modal (#3981)
# Motivation

For consistency reasons we want to rename the `TokenModal` to
`EthTokenModal`. The naming pattern should be the same for Icp, eth and
btc.

# Changes

- renames `TokenModal`

---------
