+++ src/frontend/src/eth/components/tokens/EthTokenModal.svelte
