+++ src/frontend/src/lib/derived/modal.derived.ts
@@ -93,9 +93,9 @@ export const modalBtcTransaction: Readable<boolean> = derived(
 	modalStore,
 	($modalStore) => $modalStore?.type === 'btc-transaction'
 );
-export const modalToken: Readable<boolean> = derived(
+export const modalEthToken: Readable<boolean> = derived(
 	modalStore,
-	($modalStore) => $modalStore?.type === 'token'
+	($modalStore) => $modalStore?.type === 'eth-token'
 );
 export const modalIcToken: Readable<boolean> = derived(
 	modalStore,
